+++ R/verbs_for_attributes.R
@@ -40,7 +40,7 @@
 #' df <- data.frame(x = sample(10, 5, rep = TRUE),
 #'                  y = sample(10, 5, rep = TRUE)) %>%
 #'   define(example="yes",
-#'           package="dplyr")
+#'          package="dplyr")
 #' tribe(df)
 #'
 #' take(df, names)