+++ NEWS.md
@@ -1,6 +1,6 @@
 # tribe 0.1.2
 
-* The pipe '%at>%' is renamed into '%@>%'. 
+* The pipe '%at>%' is renamed '%@>%'. 
 * Dplyr verbs for list manipulations and 'mutate_which' function 
 are transferred to the 'lplyr' package
 * Infix attribute accessor "%@%" is added (borrowed from package 'purrr')