+++ NEWS.md
@@ -5,6 +5,7 @@
 are transferred to the 'lplyr' package
 * Infix attribute accessor "%@%" is added (borrowed from package 'purrr')
 
+
 # tribe 0.1.1
 
 * Minor corrections